+++ util/massdm.py
@@ -5,7 +5,7 @@ import requests
 import json
 import Hazard
 from colorama import Fore
-from util.plugins.common import clear, setTitle, print_slow, getheaders, THIS_VERSION
+from util.plugins.common import setTitle, print_slow, getheaders
 
 def MassDM(token, Message):
     headers = {'Authorization': token}
@@ -13,8 +13,8 @@ def MassDM(token, Message):
     for channel in channelIds:
         try:
             setTitle(f"Messaging "+channel['id'])
-            requests.post(f'https://discord.com/api/v8/channels/'+channel['id']+'/messages', 
-            headers=headers, 
+            requests.post(f'https://discord.com/api/v9/channels/'+channel['id']+'/messages', 
+            headers=headers,
             data={"content": f"{Message}"})
             print(f"{Fore.RED}Messaged ID: {Fore.WHITE}"+channel['id']+Fore.RESET)
         except Exception as e: