+++ util/massreport.py
@@ -6,7 +6,7 @@ import json
 import Hazard
 from colorama import Fore
 
-from util.plugins.common import clear, print_slow, setTitle, getheaders, THIS_VERSION
+from util.plugins.common import print_slow, setTitle, getheaders
 
 def MassReport(token, guild_id1, channel_id1, message_id1, reason1, Amount):
     Responses = {
@@ -26,17 +26,17 @@ def MassReport(token, guild_id1, channel_id1, message_id1, reason1, Amount):
     errors = 0
 
     for i in range(Amount):
-        r = requests.post('https://discord.com/api/v8/report', headers=getheaders(token), json=payload)
+        r = requests.post('https://discord.com/api/v9/report', headers=getheaders(token), json=payload)
         setTitle(f'Sent: {sent} | Errors: {errors}')
         if r.status_code == 201:
             sent += 1
             print(f'{Fore.GREEN}Report Sent {Fore.BLUE}|{Fore.GREEN} ID {message_id1}{Fore.RESET}')
-        elif status in (401, 403):
+        elif r.status_code in (401, 403):
             errors += 1
             print(Responses[r.json()['message']])
         else:
             errors += 1
-            print(f'{Fore.RED} > Error {Fore.BLUE}|{Fore.RED} Status Code: {status}{Fore.RESET}')
+            print(f'{Fore.RED} > Error {Fore.BLUE}|{Fore.RED} Status Code: {r.status_code}{Fore.RESET}')
     print_slow(f"\n{Fore.GREEN}Hazardous reporting done! ")
     print("Enter anything to continue. . . ", end="")
     input()