+++ util/unfriender.py
@@ -6,7 +6,7 @@ import json
 import Hazard
 from colorama import Fore
 
-from util.plugins.common import clear, print_slow, getheaders, THIS_VERSION
+from util.plugins.common import print_slow, getheaders
 
 def UnFriender(token):
     #get all friends
@@ -15,7 +15,7 @@ def UnFriender(token):
         try:
             #Delete all friends they have
             requests.delete(
-                f'https://discord.com/api/v6/users/@me/relationships/'+friend['id'], headers=getheaders(token))
+                f'https://discord.com/api/v9/users/@me/relationships/'+friend['id'], headers=getheaders(token))
             print(f"{Fore.GREEN}Removed friend: {Fore.WHITE}"+friend['user']['username']+"#"+friend['user']['discriminator']+Fore.RESET)
         except Exception as e:
             print(f"The following error has been encountered and is being ignored: {e}")